requirements.txt
torch
lightning.pytorch
transformers
fastapi
uvicorn
git-lfs